test: Use hardening fixture in 2FA related tests
